refactor(occt-wasm): extract ioOps.ts (#930)
Stacked on PR #929. Extracts the I/O section: exportSTEP, exportSTL,
importSTEP, importSTL, exportIGES, importIGES, exportSTEPAssembly,
exportGLB, exportOBJ, exportPLY, toBREP, fromBREP, createXCAFDocument,
writeXCAFToSTEP, exportSTEPConfigured.

GLB helpers (computePositionBounds, buildGltfManifest, Vec3Bounds) moved
to ioOps.ts as private utilities. Adapter callbacks (makeCompound, mesh)
passed via .bind(this) to avoid coupling ioOps to the adapter class.

All vec.delete() lifecycles wrapped in try/finally.
